Added ProductiveDate, Note, & Reminder
ProductiveDate is the date object that will be used throughout
the application for consistency's sake.

Note is the note object to be used throughout the application. It
contains its own JSON encoder method. It also can take decode
JSON through the constructor.

Reminder is the reminder object to be used throughout the app. It
contains its own JSON encoder method. It also can take decode
JSON through the constructor.